+++ addon/helpers/build-secondary-nav-links.js
@@ -33,6 +33,10 @@ export default Ember.Helper.extend({  // Helper defined using a class, so can in
                     type: 'search'
                 },
                 {
+                    name: 'eosf.navbar.support',
+                    href: serviceLinks.osfSupport
+                },
+                {
                     name: 'eosf.navbar.donate',
                     href: 'https://cos.io/donate',
                     type: 'donateToCOS'